feature(kustomize): add integration for kustomize
implements support for kustomize as a templating engine. one of the
interesting challenges with kustomize is that it depends on files
adjacent to a single kustomization file for it's templating. the
kustomization file refers to files surrounding it. these files are
collected and layered on top of eachother to produce a single, fully
rendered manifest. because of this we've had to implement a special
handling of kustomization files. essentially, before we attempt to
render a kustomization we parse the file and generate artifact requests
for any dependencies. we'll subsequently do this for any dependencies
that happen to be kustomization files themselves.